// src/app/models/flightdetailsReturn.ts
export interface FlightDetailsReturn {
  DepartureOrigin: string;
  DepartureDestination: string;
  DepartureFare: number;
  DepartureTime: Date;
  ArrivalTime: Date;
  DepartureAirline: string;
  DepartureFlightNo: string;
  ReturnOrigin: string;
  ReturnDestination: string;
  ReturnFare: number;
  ReturnTime: Date;
  ReturnArrivalTime: Date;
  ReturnAirline: string;
  ReturnFlightNo: string;
}